git-svn-id: https://subversion.assembla.com/svn/mediasl/Projectes%20MPLAB/EOS/trunk@6330 4ad52a4b-eac5-49b1-a5a4-524d969995ed

+++ v2/Demo/FormsDemo/appMainForm.cpp
@@ -56,7 +56,7 @@ void MainForm::onSelectorPress() {
 
 
 void MainForm::onSelectorMove(
-    int position, 
+    int16_t position, 
     SelectorDirection direction) {
 }
 
@@ -88,7 +88,7 @@ void MainForm::onPaint(
 
 void MainForm::menuClickItemEventHandler(
     MenuFormHandle menuForm,
-    unsigned itemId) {
+    uint16_t itemId) {
 
     currentCommand = itemId;
