+++ v2/Include/Controllers/Display/Drivers/eosILI9341.hpp
@@ -24,31 +24,31 @@ namespace eos {
     class ILI9341_Driver: public IDisplayDriver {
         private:
             ILI9341_IO io;
-            int xScreenSize;
-            int yScreenSize;
+            int16_t xScreenSize;
+            int16_t yScreenSize;
             
         public:
             ILI9341_Driver();
             void initialize();
             void shutdown();
             void setOrientation(Orientation orientation);
-            int getXSize() const { return xScreenSize; }
-            int getYSize() const { return yScreenSize; }
+            int16_t getXSize() const { return xScreenSize; }
+            int16_t getYSize() const { return yScreenSize; }
             void clear(Color color);
-            void setPixel(int xPos, int yPos, Color color);
-            void setHPixels(int xPos, int yPos, int size, Color color);
-            void setVPixels(int xPos, int yPos, int size, Color color);
-            void setPixels(int xPos, int yPos, int xSize, int ySize, Color color);
-            void writePixels(int xPos, int yPos, int xSize, int ySize, const Color *colors);
-            void readPixels(int xPos, int yPos, int xSize, int ySize, Color *colors);
-            void vScroll(int delta, int xPos, int yPos, int xSize, int ySize);
-            void hScroll(int delta, int xPos, int yPos, int xSize, int ySize);   
+            void setPixel(int16_t xPos, int16_t yPos, Color color);
+            void setHPixels(int16_t xPos, int16_t yPos, int16_t size, Color color);
+            void setVPixels(int16_t xPos, int16_t yPos, int16_t size, Color color);
+            void setPixels(int16_t xPos, int16_t yPos, int16_t xSize, int16_t ySize, Color color);
+            void writePixels(int16_t xPos, int16_t yPos, int16_t xSize, int16_t ySize, const Color *colors);
+            void readPixels(int16_t xPos, int16_t yPos, int16_t xSize, int16_t ySize, Color *colors);
+            void vScroll(int16_t delta, int16_t xPos, int16_t yPos, int16_t xSize, int16_t ySize);
+            void hScroll(int16_t delta, int16_t xPos, int16_t yPos, int16_t xSize, int16_t ySize);   
             
         private:
-            void writePixel(Color color, unsigned count);
-            void writePixel(const Color *colors, unsigned count);
-            void readPixel(Color *colors, unsigned count);
-            void selectRegion(int x, int y, int width, int height);
+            void writePixel(Color color, int32_t count);
+            void writePixel(const Color *colors, int32_t count);
+            void readPixel(Color *colors, int32_t count);
+            void selectRegion(int16_t x, int16_t y, int16_t width, int16_t height);
             void startMemoryWrite();
             void startMemoryRead();
     };
